add byte offset

+++ src/main/java/com/garliclord/spalk/Result.java
@@ -2,5 +2,5 @@ package com.garliclord.spalk;
 
 import java.util.List;
 
-public record Result(boolean isSuccessful, List<Integer> pids, int errorPacketIndex, int errorByteOffset) {
+public record Result(boolean isSuccessful, List<Integer> pids, long errorPacketIndex, long errorByteOffset) {
 }
